打赏推文OK

+++ guns-vip-main/src/main/java/cn/stylefeng/guns/modular/note/model/params/QxPayLogParam.java
@@ -1,10 +1,10 @@
 package cn.stylefeng.guns.modular.note.model.params;
 
-import lombok.Data;
-import cn.stylefeng.roses.kernel.model.validator.BaseValidatingParam;
-import java.util.Date;
 import java.io.Serializable;
-import java.math.BigDecimal;
+import java.util.Date;
+
+import cn.stylefeng.roses.kernel.model.validator.BaseValidatingParam;
+import lombok.Data;
 
 /**
  * <p>
@@ -63,7 +63,7 @@ public class QxPayLogParam implements Serializable, BaseValidatingParam {
     /**
      * 金额
      */
-    private BigDecimal amount;
+    private Integer amount;
 
     /**
      * 付费类型 0-约单支出；1-约单汇入；2-打赏支出；3-打赏汇入；4-兑换商品支出；5-购买礼物支出；6-付费日记支出；7-付费日记汇入；8-违约金支出；9-违约金汇入；
